Added transitions after login
