+++ templates/authentication/signup.html
@@ -49,10 +49,10 @@
     </div>
     <div id="invalid-popup-signup" class="invalid-popup"></div>
 </form>
-<a class="authentication-nav nav-target" data-target="login"  href="login">
+<a class="authentication-nav nav-target" data-target="login" href="login">
     Already have an account?
     <br/>
     Login!
 </a>
-<button id="signup-button" form="signup-form" type="submit" class="authentication-button">SignUp</button>
+<button id="signup-button" form="signup-form" data-target="login" type="submit" class="authentication-button">SignUp</button>
 {% endblock%}
